+++ ml_core.lua
@@ -1,9 +1,10 @@
-﻿--[[	RCLootCouncil by Potdisc
-ml_core.lua	Contains core elements for the MasterLooter
-	-	Although possible, this module shouldn't be replaced unless closely replicated as other default modules depend on it.
-	-	Assumes several functions in SessionFrame and VotingFrame
+--[[--- ml_core.lua	Contains core elements for the MasterLooter.
+	Although possible, this module shouldn't be replaced unless closely replicated as other default modules depend on it.
+	Assumes several functions in SessionFrame and VotingFrame.
+	@author Potdisc
+]]
 
-	TODOs/NOTES:
+--[[TODOs/NOTES:
 		- SendMessage() on AddItem() to let userModules know it's safe to add to lootTable. Might have to do it other places too.
 ]]
 
@@ -56,12 +57,12 @@ end
 function RCLootCouncilML:AddItem(item, bagged, slotIndex, index)
 	addon:DebugLog("ML:AddItem", item, bagged, slotIndex, index)
 	local name, link, rarity, ilvl, iMinLevel, type, subType, iStackCount, equipLoc, texture = GetItemInfo(item)
-	local isTier = type(RCTokenClasses[addon:GetItemIDFromLink(link)]) == "table"
-	self.lootTable[index or #self.lootTable + 1] = {
+	local itemID = link and addon:GetItemIDFromLink(link)
+	self.lootTable[index or #self.lootTable + 1] = { -- We want to reserve the index even if we haven't fully loaded the item
 		["bagged"]		= bagged,
 		["lootSlot"]	= slotIndex,
 		["awarded"]		= false,
-		["name"]			= name,
+		["name"]			= name, -- REVIEW This is really not needed as it's contained in itemLink. Remove next time we break backwards com
 		["link"]			= link,
 		["quality"]		= rarity,
 		["ilvl"]			= ilvl,
@@ -69,9 +70,10 @@ function RCLootCouncilML:AddItem(item, bagged, slotIndex, index)
 		["equipLoc"]	= equipLoc,
 		["texture"]		= texture,
 		["boe"]			= addon:IsItemBoE(link),
-		["isTier"]		= isTier,
+		["relic"]		= itemID and IsArtifactRelicItem(itemID) and select(3, C_ArtifactUI.GetRelicInfoByItemID(itemID)),
+		["token"]		= itemID and RCTokenTable[itemID],
 	}
-	-- Item isn't properly loaded, so update the data in 1 sec (Should only happen with /rc test)
+		-- Item isn't properly loaded, so update the data in 1 sec (Should only happen with /rc test)
 	if not name then
 		self:ScheduleTimer("Timer", 1, "AddItem", item, bagged, slotIndex, index or #self.lootTable)
 		addon:Debug("Started timer:", "AddItem", "for", item)
@@ -353,10 +355,15 @@ function RCLootCouncilML:OnEvent(event, ...)
 end
 
 function RCLootCouncilML:LootOpened()
+	local sessionframe = addon:GetActiveModule("sessionframe")
 	if addon.isMasterLooter and GetNumLootItems() > 0 then
-		addon.target = GetUnitName("target") or L["Unknown/Chest"] -- capture the boss name
+		if addon.target and addon.target ~= "" then -- Capture boss name
+			local target = GetUnitName("target") or L["Unknown/Chest"]
+			if not (UnitInRaid(target) or UnitInParty(target)) then
+				addon.target = target -- Make sure we can't target one of our raid members for this
+			end
+		end
 		local updatedLootSlot = {}
-		self.lootTable = {} -- Clear it just in case it was populated.
 		for i = 1, GetNumLootItems() do
 			local item = GetLootSlotLink(i)
 			-- We have reopened the loot frame, so check if we should update .lootSlot
@@ -376,15 +383,19 @@ function RCLootCouncilML:LootOpened()
 				end
 			else
 				if db.altClickLooting then self:ScheduleTimer("HookLootButton", 0.5, i) end -- Delay lootbutton hooking to ensure other addons have had time to build their frames
-				local _, _, quantity, quality = GetLootSlotInfo(i)
-				if self:ShouldAutoAward(item, quality) and quantity > 0 then
-					self:AutoAward(i, item, quality, db.autoAwardTo, db.autoAwardReason, addon.target)
+				-- We might already have the session frame running, in which case we shouldn't add the items again
+				if not sessionframe:IsRunning() then
 
-				elseif self:CanWeLootItem(item, quality) and quantity > 0 then -- check if our options allows us to loot it
-					self:AddItem(item, false, i)
+					local _, _, quantity, quality = GetLootSlotInfo(i)
+					if self:ShouldAutoAward(item, quality) and quantity > 0 then
+						self:AutoAward(i, item, quality, db.autoAwardTo, db.autoAwardReason, addon.target)
 
-				elseif quantity == 0 then -- it's coin, just loot it
-					LootSlot(i)
+					elseif self:CanWeLootItem(item, quality) and quantity > 0 then -- check if our options allows us to loot it
+						self:AddItem(item, false, i)
+
+					elseif quantity == 0 then -- it's coin, just loot it
+						LootSlot(i)
+					end
 				end
 			end
 		end
@@ -606,11 +617,11 @@ function RCLootCouncilML:AutoAward(lootIndex, item, quality, name, reason, boss)
 end
 
 local history_table = {}
-function RCLootCouncilML:TrackAndLogLoot(name, item, response, boss, votes, itemReplaced1, itemReplaced2, reason)
+function RCLootCouncilML:TrackAndLogLoot(name, item, response, boss, votes, itemReplaced1, itemReplaced2, reason, isToken)
 	if reason and not reason.log then return end -- Reason says don't log
 	if not (db.sendHistory or db.enableHistory) then return end -- No reason to do stuff when we won't use it
 	if addon.testMode and not addon.nnp then return end -- We shouldn't track testing awards.
-	local instanceName, _, _, difficultyName = GetInstanceInfo()
+	local instanceName, _, difficultyID, difficultyName, _,_,_,mapID, groupSize = GetInstanceInfo()
 	addon:Debug("ML:TrackAndLogLoot()")
 	history_table["lootWon"] 		= item
 	history_table["date"] 			= date("%d/%m/%y")
@@ -624,7 +635,11 @@ function RCLootCouncilML:TrackAndLogLoot(name, item, response, boss, votes, item
 	history_table["responseID"] 	= response or reason.sort - 400 										-- Changed in v2.0 (reason responseID was 0 pre v2.0)
 	history_table["color"]			= reason and reason.color or db.responses[response].color	-- New in v2.0
 	history_table["class"]			= self.candidates[name].class											-- New in v2.0
-	history_table["isAwardReason"] = reason and true or false											-- New in v2.0
+	history_table["isAwardReason"]= reason and true or false												-- New in v2.0
+	history_table["difficultyID"]	= difficultyID																-- New in v2.3+
+	history_table["mapID"]			= mapID																		-- New in v2.3+
+	history_table["groupSize"]		= groupSize																	-- New in v2.3+
+	history_table["tierToken"]		= isToken																	-- New in v2.3+
 
 	if db.sendHistory then -- Send it, and let comms handle the logging
 		addon:SendCommand("group", "history", name, history_table)
@@ -777,9 +792,10 @@ LibDialog:Register("RCLOOTCOUNCIL_CONFIRM_AWARD", {
 				-- IDEA Perhaps come up with a better way of handling this
 				local session, player, response, reason, votes, item1, item2 = unpack(data,1,7)
 				local item = RCLootCouncilML.lootTable[session].link -- Store it now as we wipe lootTable after Award()
+				local isToken = RCLootCouncilML.lootTable[session].token
 				local awarded = RCLootCouncilML:Award(session, player, response, reason)
 				if awarded then -- log it
-					RCLootCouncilML:TrackAndLogLoot(player, item, response, addon.target, votes, item1, item2, reason)
+					RCLootCouncilML:TrackAndLogLoot(player, item, response, addon.target, votes, item1, item2, reason, isToken)
 				end
 				-- We need to delay the test mode disabling so comms have a chance to be send first!
 				if addon.testMode and RCLootCouncilML:HasAllItemsBeenAwarded() then RCLootCouncilML:EndSession() end